gh-pages updated

+++ deploy_dist.sh
@@ -1,5 +1,11 @@
 shopt -s extglob
-rm -r !(deploy_dist.sh|dist|node_modules)
-mv dist/* .
-rm -r dist
+
+if [ -d "dist" ]; then
+  rm -r !(deploy_dist.sh|dist|node_modules)
+  mv dist/* .
+  rm -r dist
+  git add *
+  git commit -am "gh-pages updated"
+  git push
+fi
